[53] - Coin refactor
