+++ app/Infrastructure/Providers/AppServiceProvider.php
@@ -5,7 +5,7 @@ namespace App\Infrastructure\Providers;
 use App\Domain\DataSources\CoinDataSource;
 use App\Domain\DataSources\UserDataSource;
 use App\Domain\DataSources\WalletDataSource;
-use App\Infrastructure\ApiServices\CoinloreApiService;
+use App\Infrastructure\ApiClients\CoinloreApiClient;
 use App\Infrastructure\Persistence\ApiCoinDataSource;
 use App\Infrastructure\Persistence\FileUserDataSource;
 use App\Infrastructure\Persistence\CacheWalletDataSource;
@@ -34,7 +34,7 @@ class AppServiceProvider extends ServiceProvider
             return new FileUserDataSource();
         });
         $this->app->bind(CoinDataSource::class, function () {
-            return new ApiCoinDataSource(new CoinloreApiService());
+            return new ApiCoinDataSource(new CoinloreApiClient());
         });
         $this->app->bind(WalletDataSource::class, function () {
             return new CacheWalletDataSource();
